Changes done in AStart related files

+++ code/src/dk/dtu/ai/blueducks/heuristics/GoToBoxHeuristic.java
@@ -5,10 +5,10 @@ import dk.dtu.ai.blueducks.map.State;
 
 public class GoToBoxHeuristic implements Heuristic<State, GoToBoxGoal> {
 
-	private double distance;
+	private float distance;
 
 	@Override
-	public double getHeuristicValue(State state, GoToBoxGoal goal) {
+	public float getHeuristicValue(State state, GoToBoxGoal goal) {
 		distance = Math.abs(goal.getTo().x - state.getAgentAt().x)
 				+ Math.abs(goal.getTo().y - state.getAgentAt().y);
 		return distance;
